Brayan

+++ src/app/components/list-catalogue/list-catalogue.component.ts
@@ -1,6 +1,7 @@
 import { Component, NgModule, OnInit } from '@angular/core';
 import {  NgxSpinnerService } from 'ngx-spinner';
 import { CatalogueService } from 'src/app/services/catalogue/catalogue.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-list-catalogue',
@@ -28,19 +29,41 @@ export class ListCatalogueComponent implements OnInit {
   }
 
   deleter(id:Number): void{
-    this.spinner2.show();
-    this.imagenService.deleteCatalogue(id).subscribe(
-      data=>{
-        if(data){
-          this.spinner2.hide();
-          this.loaderImage()
-        }
-      },err=>{
-        this.spinner2.hide();;
-        alert(err);
-        console.log(err);
+    Swal.fire({
+      title: 'Eliminar catálogo',
+      text: "¿Desea eliminar el catálogo?",
+      icon: 'question',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Aceptar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.spinner2.show();
+        this.imagenService.deleteCatalogue(id).subscribe(
+          data=>{
+            if(data){
+              this.spinner2.hide();
+              this.loaderImage()
+            }
+          },err=>{
+            this.spinner2.hide();;
+            alert(err);
+            console.log(err);
+          }
+        )
+        Swal.fire(
+          'Catálogo eliminado',
+          `Se ha eliminar el catálogo con exito.`,
+          'success'
+        )
+        setTimeout(() => {
+      
+          window.location.reload()
+        }, 1000);
       }
-    )
+    })
+
   }
 
   openModal(){
